rename toFirehoseTester -> firehoseTester
