+++ audit/audit-template.yml
@@ -486,10 +486,10 @@ Resources:
         EntryPoints:
           - delimiter-app.ts
 
-  ToFirehoseTesterLambdaAccessRole:
+  firehoseTesterLambdaAccessRole:
     Type: AWS::IAM::Role
     Properties:
-      RoleName: !Sub "${AWS::StackName}-LambdaAccessRole-ToFirehoseTester"
+      RoleName: !Sub "${AWS::StackName}-LambdaAccessRole-firehoseTester"
       PermissionsBoundary: !If
         - UsePermissionsBoundary
         - !Ref PermissionsBoundary
@@ -506,19 +506,19 @@ Resources:
       ManagedPolicyArns:
         - "arn:aws:iam::aws:policy/service-role/AWSLambdaKinesisExecutionRole"
 
-  ToFirehoseTesterFunction:
+  firehoseTesterFunction:
     DependsOn:
-      - ToFirehoseTesterLambdaAccessRole
+      - firehoseTesterLambdaAccessRole
     Type: "AWS::Serverless::Function"
     Condition: IsTestStage
     Properties:
-      FunctionName: "ToFirehoseTester"
+      FunctionName: "firehoseTester"
       CodeUri: ../audit/lambda
       PackageType: Zip
-      Handler: toFirehoseTester-app.handler
+      Handler: firehoseTester-app.handler
       Runtime: nodejs14.x
       Timeout: 30
-      Role: !GetAtt toFirehoseTesterLambdaAccessRole.Arn
+      Role: !GetAtt firehoseTesterLambdaAccessRole.Arn
       KmsKeyArn: !GetAtt LambdaKMSKey.Arn
     Metadata: # Manage esbuild properties
       BuildMethod: esbuild
@@ -527,14 +527,14 @@ Resources:
         Target: "es2020"
         Sourcemap: true
         EntryPoints:
-          - toFirehoseTester-app.ts
+          - firehoseTester-app.ts
 
-  ToFirehoseTesterInvokePermission:
+  firehoseTesterInvokePermission:
     Type: AWS::Lambda::Permission
     Condition: DeployTestPolicy
     Properties:
       Action: lambda:InvokeFunction
-      FunctionName: !Ref ToFirehoseTesterFunction
+      FunctionName: !Ref firehoseTesterFunction
       Principal: !Ref TestRoleArn
 
   #
