+++ audit/lambda/firehoseTester-app.ts
